+++ sqflite_common/CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 2.3.0
 
-- Add `startTransaction` parameter to `batch()` to control whether sqflite will
-  start a transaction for this batch or not.
+- Add `apply()` method to `Batch`. It will execute statements in that batch
+  without starting a new transaction.
 
 ## 2.2.1+1